Support anthropic extended thinking
Adds support for anthropic extended thinking.

Messages are now parsed into content objects with a list of blocks,
instead of a string for content. This is a common openai based format,
so existing providers work as normal.

Thinking / redacted thinking blocks are represented as blocks within
markup tags. The signature is a property as a closing tag, which is a
bit odd. This was done to make it easy to support streaming (the
signature comes at the end of the message).

Redacted thinking is also supported. Redacted thinking doesn't have a
signature (because the content is encrypted, so it's already tamper
proof).